requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
openai
tiktoken
pandas
pdfplumber
scikit-learn
plotly
python-docx
pywin32
